add open faas deployment and fix tests run
