--- tmf-service/.vscode/launch.json
@@ -1,19 +1,0 @@
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "type": "node",
-            "request": "launch",
-            "name": "Launch TMS Engine",
-            "skipFiles": [
-                "<node_internals>\\**"
-            ],
-            "program": "${workspaceFolder}\\src\\index.ts",
-            "preLaunchTask": "tsc: build - tsconfig.json",
-            "outFiles": [
-                "${workspaceFolder}\\build\\**\\*.js"
-            ],
-            "cwd": "${workspaceFolder}"
-        }
-    ]
-}
