--- tmf-service/docker-compose.yaml
@@ -1,11 +1,0 @@
-version: "2"
-services:
-  redis:
-    image: redis
-    ports:
-      - "6379:6379"
-  node:
-    image: tms-api
-    restart: "no"
-    depends_on:
-      - "redis"
