--- tmf-service/dockerfile
@@ -1,21 +1,0 @@
-FROM node:12.16.1-alpine as builder
-
-WORKDIR /opt/tms-api
-
-RUN apk add --no-cache -t build-dependencies git make gcc g++ python libtool autoconf automake \
-    && cd $(npm root -g)/npm \
-    && npm config set unsafe-perm true \
-    && npm install -g node-gyp
-
-COPY package.json package-lock.json* /opt/tms-api/
-
-RUN npm ci
-
-RUN apk del build-dependencies
-
-COPY tsconfig.json /opt/tms-api/
-COPY src /opt/tms-api/src
-
-RUN npm run build
-
-CMD ["node", "build/index.js"]
