+++ tms-service/docker-compose.yaml
@@ -1,0 +1,5 @@
+version: "2"
+services:
+  node:
+    image: tms-service
+    restart: "no"
