+++ tms-service/src/controllers/misc.ts
@@ -2,7 +2,6 @@ import { Context } from 'koa';
 
 const monitorTransaction = (ctx: Context): Context => {
   ctx.body = { result: 'Transaction is valid' };
-
   return ctx;
 };
 
@@ -10,7 +9,6 @@ const healthcheck = (ctx: Context): Context => {
   const data = {
     status: 'UP',
   };
-
   ctx.body = data;
 
   return ctx;
