+++ tms-service/src/routes.ts
@@ -3,8 +3,8 @@ import * as miscController from './controllers/misc';
 
 const router = new Router();
 
+router.get('/', miscController.healthcheck);
 router.get('/health', miscController.healthcheck);
-
 router.post('/monitor/transaction', miscController.monitorTransaction);
 
 export default router;
