+++ tms-service/tsconfig.json
@@ -15,5 +15,7 @@
     "resolveJsonModule": true /* Include modules imported with '.json' extension */,
     "skipLibCheck": true /* Skip type checking of declaration files. */,
     "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
-  }
+  },
+  "include": ["src/**/*"],
+  "exclude": ["node_modules"]
 }
